requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
websockets>=10.0
pydantic>=1.8.2
aiofiles>=0.7.0
asyncio>=3.4.3

# Audio processing
pydub>=0.25.1

# Additional utilities
python-multipart>=0.0.5
